feat(zdtp-migration): relocate token manifest data arrays to src/config
Move SPACING_TOKENS, FONT_TOKENS, SIZE_TOKENS, COLOR_TOKENS out of the
legacy panel component directory into src/config/design-tokens-manifest.ts
so they survive the W3-2 deletion of the panel component tree.

The legacy manifest.ts is updated to re-export the arrays from the new
location (with a cast to the local narrower TokenDef type so existing
tab components remain type-safe). All 35 existing token-manifest tests
pass unchanged.
